// data/github/current/ui/packages/memex/src/client/api/memex-items/paginated-views.ts
export type GroupMetadata =
  | ExtendedRepository
  | IAssignee
  | IssueType
  | Iteration
  | Label
  | Milestone
  | PersistedOption
  | ParentIssue

export type SliceMetadata = GroupMetadata

export type PaginatedTotalCount = {
  value: number
  isApproximate: boolean
}

export type FieldMetric = {
  fieldId: MemexProjectColumnId
  value: number
}

export type SliceData = {
  sliceId: string
  sliceValue: string
  sliceMetadata?: SliceMetadata
  totalCount: PaginatedTotalCount
}

export type PageInfo = {
  hasNextPage: boolean
  hasPreviousPage: boolean
  endCursor?: string
  startCursor?: string
}

export type GroupedMemexItems<TMemexItemType> = {
  groupId: string
  secondaryGroupId?: string
  nodes: Array<TMemexItemType>
  pageInfo: PageInfo
}

export type Group = {
  groupId: string
  groupValue: string
  groupMetadata?: GroupMetadata
  totalCount: PaginatedTotalCount
  fieldMetrics: Array<FieldMetric>
}

export type PaginatedGroups = {
  nodes: Array<Group>
  pageInfo: PageInfo
}

type BasePaginatedData = {
  totalCount: PaginatedTotalCount
  slices?: Array<SliceData>
  exception?: string
}

export type PaginatedItemsData = BasePaginatedData & {
  nodes: Array<MemexItem>
  pageInfo: PageInfo
}

export type PaginatedGroupsData = BasePaginatedData & {
  groups: PaginatedGroups
  groupedItems: Array<GroupedMemexItems<MemexItem>>
  secondaryGroups: null
}

export type PaginatedGroupsAndSecondaryGroupsData = Omit<PaginatedGroupsData, 'secondaryGroups'> & {
  secondaryGroups: PaginatedGroups
}